add error check for dpol controller
